Import function version 2
Added the missing procedure of getting the info of a file (just on the outer part, still haven't gone into dealing with each info yet).

And more importantly, added the ability to import a whole folder into the volume. A tough challenge, if I may say.

But it hasn't been tested yet, so it'll either become an important breakthrough, or turn out to be just a time-consuming mess, with full of bugs...